backend/requirements.txt
uvicorn
fastapi
httpx
redis
pydantic
pydantic-settings
pytest
python-okx
python-dotenv
websockets
pandas